requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
cloudpickle==3.0.0
easydict==1.13
flatbuffers==2.0.7
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
joblib==1.4.2
keras==2.7.0
Keras-Preprocessing==1.1.2
lazy_loader==0.4
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.3.0
progressbar2==4.4.2
protobuf==3.20.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
python-utils==3.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.15.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.34.0
tensorlayerx==0.5.8
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.2
